pyproject.toml
[project]
name = "fodmap-app"
version = "0.1.0"
description = "FODMAP community social media app"
authors = [
    {name = "Paul McJannet"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==5.2.1)",
    "djangorestframework (==3.16.0)",
    "drf-nested-routers (>=0.94.2,<0.95.0)",
    "asgiref (==3.8.1)",
    "django-cors-headers (==4.7.0)",
    "djangorestframework-simplejwt (==5.5.0)",
    "pillow (==11.2.1)",
    "pyjwt (==2.9.0)",
    "python-decouple (==3.8)",
    "python-dotenv (==1.1.1)",
    "sqlparse (==0.5.3)",
    "tzdata (==2025.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
